Refactor category image views
